Rename 'state' to 'reducers'
