+++ client/reducers/domain-search/index.js
